refactor(step3): 매장 주문과 매장 테이블 aggregation 분리
