+++ src/test/java/kitchenpos/eatinorders/tobe/domain/EatInOrderLineItemTest.java
@@ -1,5 +1,7 @@
 package kitchenpos.eatinorders.tobe.domain;
 
+import kitchenpos.eatinorders.tobe.domain.order.EatInOrderLineItem;
+import kitchenpos.eatinorders.tobe.domain.order.EatInOrderLineItemMenu;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
