--- src/test/java/kitchenpos/eatinorders/tobe/domain/OrderDateTime.java
@@ -1,4 +1,0 @@
-package kitchenpos.eatinorders.tobe.domain;
-
-public class OrderDateTime {
-}
