--- src/test/java/kitchenpos/eatinorders/tobe/domain/OrderStatus.java
@@ -1,4 +1,0 @@
-package kitchenpos.eatinorders.tobe.domain;
-
-public enum OrderStatus {
-}
