+++ src/test/java/kitchenpos/eatinorders/tobe/domain/order/EatInOrder.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe.domain;
+package kitchenpos.eatinorders.tobe.domain.order;
 
 import java.util.Objects;
 import java.util.UUID;
