+++ src/test/java/kitchenpos/eatinorders/tobe/domain/order/EatInOrderLineItemMenu.java
@@ -1,6 +1,6 @@
-package kitchenpos.eatinorders.tobe.domain;
+package kitchenpos.eatinorders.tobe.domain.order;
 
-import kitchenpos.eatinorders.tobe.domain.vo.EatInOrderLineItemMenuPrice;
+import kitchenpos.eatinorders.tobe.domain.order.vo.EatInOrderLineItemMenuPrice;
 
 import java.util.UUID;
