+++ src/test/java/kitchenpos/eatinorders/tobe/domain/order/EatInOrderStatus.java
@@ -1,0 +1,4 @@
+package kitchenpos.eatinorders.tobe.domain.order;
+
+public class EatInOrderStatus {
+}
