+++ src/test/java/kitchenpos/eatinorders/tobe/domain/order/EatInOrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe.domain;
+package kitchenpos.eatinorders.tobe.domain.order;
 
 import java.util.UUID;
