+++ src/test/java/kitchenpos/eatinorders/tobe/domain/order/vo/EatInOrderLineItemMenuPrice.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe.domain.vo;
+package kitchenpos.eatinorders.tobe.domain.order.vo;
 
 public class EatInOrderLineItemMenuPrice {
     private final int value;
