+++ src/test/java/kitchenpos/eatinorders/tobe/domain/ordertable/OrderTable.java
@@ -1,9 +1,9 @@
-package kitchenpos.eatinorders.tobe.domain;
+package kitchenpos.eatinorders.tobe.domain.ordertable;
 
-import kitchenpos.eatinorders.tobe.domain.vo.NumberOfGuests;
-import kitchenpos.eatinorders.tobe.domain.vo.Occupied;
-import kitchenpos.eatinorders.tobe.domain.vo.OrderTableId;
-import kitchenpos.eatinorders.tobe.domain.vo.OrderTableName;
+import kitchenpos.eatinorders.tobe.domain.ordertable.vo.NumberOfGuests;
+import kitchenpos.eatinorders.tobe.domain.ordertable.vo.Occupied;
+import kitchenpos.eatinorders.tobe.domain.ordertable.vo.OrderTableId;
+import kitchenpos.eatinorders.tobe.domain.ordertable.vo.OrderTableName;
 
 import java.util.UUID;
