+++ src/test/java/kitchenpos/eatinorders/tobe/domain/ordertable/vo/OrderTableName.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.tobe.domain.vo;
+package kitchenpos.eatinorders.tobe.domain.ordertable.vo;
 
 import java.util.Objects;
